ml-embedded-evaluation-kit-refs_tags_21.05/model_conditioning_examples/requirements.txt
tensorflow-gpu==2.3.4
tensorflow-model-optimization==0.5.0
numpy==1.18.5